Tetrazone: fixed the extract_digits() function
  * curval was too small to fit the intermediate value,
    extended to UINT16 fixed it
  * more memory for samples
  * added insert_hiscore() function with tests

+++ src/tetris/utils.c
@@ -20,7 +20,7 @@ UINT8 extract_digits(UINT8 *digit_buffer, UINT8 max_digits, UINT16 value)
 {
     UINT8 index = 0;
     UINT8 out_digit = value % 10;
-    UINT8 cur_value = value / 10;
+    UINT16 cur_value = value / 10;
     digit_buffer[index++] = '0' + out_digit;
     while (cur_value > 0) {
         out_digit = cur_value % 10;
